+++ src/renderer/services/midgard/types.ts
@@ -222,6 +222,8 @@ export type PoolShareType = DepositType | 'all'
 export type PoolShare = {
   units: BigNumber
   asset: Asset
+  assetAddress: O.Option<Address>
+  runeAddress: O.Option<Address>
   assetAddedAmount: BaseAmount
   type: PoolShareType
 }